Crate aws_sdk_sts

Security Token Service

Security Token Service (STS) enables you to request temporary, limited-privilege credentials for Identity and Access Management (IAM) users or for users that you authenticate (federated users). This guide provides descriptions of the STS API. For more information about using this service, see Temporary Security Credentials.

Crate Organization

The entry point for most customers will be Client. Client exposes one method for each API offered by the service.

Some APIs require complex or nested arguments. These exist in model.

Lastly, errors that can be returned by the service are contained within error. Error defines a meta error encompassing all possible errors that can be returned by the service.

The other modules within this crate are not required for normal usage.

Modules

client

Client and fluent builders for calling the service.

config

Configuration for the service.

error

Errors that can occur when calling the service.

input

Input structures for operations.

middleware

Base Middleware Stack

model

Data structures used by operation inputs/outputs.

operation

All operations that this crate can perform.

output

Output structures for operations.

types

Re-exported types from supporting crates.

Structs

AppName

App name that can be configured with an AWS SDK client to become part of the user agent string.

Client

Client for AWS Security Token Service

Config

Service config.

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

RetryConfig

Retry configuration for requests.

Enums

Error

All possible error types for this service.

Statics

PKG_VERSION

Crate version number.