aws_sdk_sts

Module client

Client for calling AWS Security Token Service.

Using the Client

A client has a function for every operation that can be performed by the service. For example, the AssumeRole operation has a Client::assume_role, function which returns a builder for that operation. The fluent builder ultimately has a send() function that returns an async future that returns a result, as illustrated below:

let result = client.assume_role()
    .role_arn("example")
    .send()
    .await;

The underlying HTTP requests that get made by this can be modified with the customize_operation function on the fluent builder. See the customize module for more information.

Modules

customize

Operation customization and supporting types.

Structs

Client

Client for AWS Security Token Service