Trait aws_sdk_sts::config::ProvideCredentials

pub trait ProvideCredentials: Send + Sync + Debug {
    // Required method
    fn provide_credentials<'a>(&'a self) -> ProvideCredentials<'a>
       where Self: 'a;

    // Provided method
    fn fallback_on_interrupt(&self) -> Option<Credentials> { ... }
}

Asynchronous Credentials Provider

Required Methods

fn provide_credentials<'a>(&'a self) -> ProvideCredentials<'a>
where Self: 'a,

Returns a future that provides credentials.

Provided Methods

fn fallback_on_interrupt(&self) -> Option<Credentials>

Returns fallback credentials.

This method should be used as a fallback plan, i.e., when a call to provide_credentials is interrupted and its future fails to complete.

The fallback credentials should be set aside and ready to be returned immediately. Therefore, the user should NOT go fetch new credentials within this method, which might cause a long-running operation.