Module aws_sdk_sts::error

Modules

expired_token_exception

See ExpiredTokenException

idp_communication_error_exception

See IdpCommunicationErrorException

idp_rejected_claim_exception

See IdpRejectedClaimException

invalid_authorization_message_exception

See InvalidAuthorizationMessageException

invalid_identity_token_exception

See InvalidIdentityTokenException

malformed_policy_document_exception

See MalformedPolicyDocumentException

packed_policy_too_large_exception

See PackedPolicyTooLargeException

region_disabled_exception

See RegionDisabledException

Structs

AssumeRoleError

AssumeRoleWithSAMLError

AssumeRoleWithWebIdentityError

DecodeAuthorizationMessageError

ExpiredTokenException

The web identity token that was passed is expired or is not valid. Get a new identity

GetAccessKeyInfoError

GetCallerIdentityError

GetFederationTokenError

GetSessionTokenError

IdpCommunicationErrorException

The request could not be fulfilled because the identity provider (IDP) that

IdpRejectedClaimException

The identity provider (IdP) reported that authentication failed. This might be because

InvalidAuthorizationMessageException

The error returned if the message passed to DecodeAuthorizationMessage

InvalidIdentityTokenException

The web identity token that was passed could not be validated by AWS. Get a new

MalformedPolicyDocumentException

The request was rejected because the policy document was malformed. The error message

PackedPolicyTooLargeException

The request was rejected because the total packed size of the session policies and

RegionDisabledException

STS is not activated in the requested region for the account that is being asked to

Enums

AssumeRoleErrorKind

AssumeRoleWithSAMLErrorKind

AssumeRoleWithWebIdentityErrorKind

DecodeAuthorizationMessageErrorKind

GetAccessKeyInfoErrorKind

GetCallerIdentityErrorKind

GetFederationTokenErrorKind

GetSessionTokenErrorKind