Struct aws_sdk_sts::operation::assume_role_with_saml::builders::AssumeRoleWithSamlOutputBuilder

#[non_exhaustive]
pub struct AssumeRoleWithSamlOutputBuilder { /* private fields */ }

A builder for AssumeRoleWithSamlOutput.

Implementations

impl AssumeRoleWithSamlOutputBuilder

pub fn credentials(self, input: Credentials) -> Self

The temporary security credentials, which include an access key ID, a secret access key, and a security (or session) token.

The size of the security token that STS API operations return is not fixed. We strongly recommend that you make no assumptions about the maximum size.

pub fn set_credentials(self, input: Option<Credentials>) -> Self

The temporary security credentials, which include an access key ID, a secret access key, and a security (or session) token.

The size of the security token that STS API operations return is not fixed. We strongly recommend that you make no assumptions about the maximum size.

pub fn get_credentials(&self) -> &Option<Credentials>

The temporary security credentials, which include an access key ID, a secret access key, and a security (or session) token.

The size of the security token that STS API operations return is not fixed. We strongly recommend that you make no assumptions about the maximum size.

pub fn assumed_role_user(self, input: AssumedRoleUser) -> Self

The identifiers for the temporary security credentials that the operation returns.

pub fn set_assumed_role_user(self, input: Option<AssumedRoleUser>) -> Self

The identifiers for the temporary security credentials that the operation returns.

pub fn get_assumed_role_user(&self) -> &Option<AssumedRoleUser>

The identifiers for the temporary security credentials that the operation returns.

pub fn packed_policy_size(self, input: i32) -> Self

A percentage value that indicates the packed size of the session policies and session tags combined passed in the request. The request fails if the packed size is greater than 100 percent, which means the policies and tags exceeded the allowed space.

pub fn set_packed_policy_size(self, input: Option<i32>) -> Self

A percentage value that indicates the packed size of the session policies and session tags combined passed in the request. The request fails if the packed size is greater than 100 percent, which means the policies and tags exceeded the allowed space.

pub fn get_packed_policy_size(&self) -> &Option<i32>

A percentage value that indicates the packed size of the session policies and session tags combined passed in the request. The request fails if the packed size is greater than 100 percent, which means the policies and tags exceeded the allowed space.

pub fn subject(self, input: impl Into<String>) -> Self

The value of the NameID element in the Subject element of the SAML assertion.

pub fn set_subject(self, input: Option<String>) -> Self

The value of the NameID element in the Subject element of the SAML assertion.

pub fn get_subject(&self) -> &Option<String>

The value of the NameID element in the Subject element of the SAML assertion.

pub fn subject_type(self, input: impl Into<String>) -> Self

The format of the name ID, as defined by the Format attribute in the NameID element of the SAML assertion. Typical examples of the format are transient or persistent.

If the format includes the prefix urn:oasis:names:tc:SAML:2.0:nameid-format, that prefix is removed. For example, urn:oasis:names:tc:SAML:2.0:nameid-format:transient is returned as transient. If the format includes any other prefix, the format is returned with no modifications.

pub fn set_subject_type(self, input: Option<String>) -> Self

The format of the name ID, as defined by the Format attribute in the NameID element of the SAML assertion. Typical examples of the format are transient or persistent.

If the format includes the prefix urn:oasis:names:tc:SAML:2.0:nameid-format, that prefix is removed. For example, urn:oasis:names:tc:SAML:2.0:nameid-format:transient is returned as transient. If the format includes any other prefix, the format is returned with no modifications.

pub fn get_subject_type(&self) -> &Option<String>

The format of the name ID, as defined by the Format attribute in the NameID element of the SAML assertion. Typical examples of the format are transient or persistent.

If the format includes the prefix urn:oasis:names:tc:SAML:2.0:nameid-format, that prefix is removed. For example, urn:oasis:names:tc:SAML:2.0:nameid-format:transient is returned as transient. If the format includes any other prefix, the format is returned with no modifications.

pub fn issuer(self, input: impl Into<String>) -> Self

The value of the Issuer element of the SAML assertion.

pub fn set_issuer(self, input: Option<String>) -> Self

The value of the Issuer element of the SAML assertion.

pub fn get_issuer(&self) -> &Option<String>

The value of the Issuer element of the SAML assertion.

pub fn audience(self, input: impl Into<String>) -> Self

The value of the Recipient attribute of the SubjectConfirmationData element of the SAML assertion.

pub fn set_audience(self, input: Option<String>) -> Self

The value of the Recipient attribute of the SubjectConfirmationData element of the SAML assertion.

pub fn get_audience(&self) -> &Option<String>

The value of the Recipient attribute of the SubjectConfirmationData element of the SAML assertion.

pub fn name_qualifier(self, input: impl Into<String>) -> Self

A hash value based on the concatenation of the following:

The Issuer response value.

The Amazon Web Services account ID.

The friendly name (the last part of the ARN) of the SAML provider in IAM.

The combination of NameQualifier and Subject can be used to uniquely identify a user.

The following pseudocode shows how the hash value is calculated:

BASE64 ( SHA1 ( "https://example.com/saml" + "123456789012" + "/MySAMLIdP" ) )

pub fn set_name_qualifier(self, input: Option<String>) -> Self

A hash value based on the concatenation of the following:

The Issuer response value.

The Amazon Web Services account ID.

The friendly name (the last part of the ARN) of the SAML provider in IAM.

The combination of NameQualifier and Subject can be used to uniquely identify a user.

The following pseudocode shows how the hash value is calculated:

BASE64 ( SHA1 ( "https://example.com/saml" + "123456789012" + "/MySAMLIdP" ) )

pub fn get_name_qualifier(&self) -> &Option<String>

A hash value based on the concatenation of the following:

The Issuer response value.

The Amazon Web Services account ID.

The friendly name (the last part of the ARN) of the SAML provider in IAM.

The combination of NameQualifier and Subject can be used to uniquely identify a user.

The following pseudocode shows how the hash value is calculated:

BASE64 ( SHA1 ( "https://example.com/saml" + "123456789012" + "/MySAMLIdP" ) )

pub fn source_identity(self, input: impl Into<String>) -> Self

The value in the SourceIdentity attribute in the SAML assertion.

You can require users to set a source identity value when they assume a role. You do this by using the sts:SourceIdentity condition key in a role trust policy. That way, actions that are taken with the role are associated with that user. After the source identity is set, the value cannot be changed. It is present in the request for all actions that are taken by the role and persists across chained role sessions. You can configure your SAML identity provider to use an attribute associated with your users, like user name or email, as the source identity when calling AssumeRoleWithSAML. You do this by adding an attribute to the SAML assertion. For more information about using source identity, see Monitor and control actions taken with assumed roles in the IAM User Guide.

The regex used to validate this parameter is a string of characters consisting of upper- and lower-case alphanumeric characters with no spaces. You can also include underscores or any of the following characters: =,.@-

pub fn set_source_identity(self, input: Option<String>) -> Self

The value in the SourceIdentity attribute in the SAML assertion.

You can require users to set a source identity value when they assume a role. You do this by using the sts:SourceIdentity condition key in a role trust policy. That way, actions that are taken with the role are associated with that user. After the source identity is set, the value cannot be changed. It is present in the request for all actions that are taken by the role and persists across chained role sessions. You can configure your SAML identity provider to use an attribute associated with your users, like user name or email, as the source identity when calling AssumeRoleWithSAML. You do this by adding an attribute to the SAML assertion. For more information about using source identity, see Monitor and control actions taken with assumed roles in the IAM User Guide.

The regex used to validate this parameter is a string of characters consisting of upper- and lower-case alphanumeric characters with no spaces. You can also include underscores or any of the following characters: =,.@-

pub fn get_source_identity(&self) -> &Option<String>

The value in the SourceIdentity attribute in the SAML assertion.

You can require users to set a source identity value when they assume a role. You do this by using the sts:SourceIdentity condition key in a role trust policy. That way, actions that are taken with the role are associated with that user. After the source identity is set, the value cannot be changed. It is present in the request for all actions that are taken by the role and persists across chained role sessions. You can configure your SAML identity provider to use an attribute associated with your users, like user name or email, as the source identity when calling AssumeRoleWithSAML. You do this by adding an attribute to the SAML assertion. For more information about using source identity, see Monitor and control actions taken with assumed roles in the IAM User Guide.

The regex used to validate this parameter is a string of characters consisting of upper- and lower-case alphanumeric characters with no spaces. You can also include underscores or any of the following characters: =,.@-

pub fn build(self) -> AssumeRoleWithSamlOutput

Consumes the builder and constructs a AssumeRoleWithSamlOutput.