Enum aws_sdk_sts::operation::assume_role_with_saml::AssumeRoleWithSAMLError

#[non_exhaustive]
pub enum AssumeRoleWithSAMLError {
    ExpiredTokenException(ExpiredTokenException),
    IdpRejectedClaimException(IdpRejectedClaimException),
    InvalidIdentityTokenException(InvalidIdentityTokenException),
    MalformedPolicyDocumentException(MalformedPolicyDocumentException),
    PackedPolicyTooLargeException(PackedPolicyTooLargeException),
    RegionDisabledException(RegionDisabledException),
    Unhandled(Unhandled),
}

Error type for the AssumeRoleWithSAMLError operation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ExpiredTokenException(ExpiredTokenException)

The web identity token that was passed is expired or is not valid. Get a new identity token from the identity provider and then retry the request.

IdpRejectedClaimException(IdpRejectedClaimException)

The identity provider (IdP) reported that authentication failed. This might be because the claim is invalid.

If this error is returned for the AssumeRoleWithWebIdentity operation, it can also mean that the claim has expired or has been explicitly revoked.

InvalidIdentityTokenException(InvalidIdentityTokenException)

The web identity token that was passed could not be validated by Amazon Web Services. Get a new identity token from the identity provider and then retry the request.

MalformedPolicyDocumentException(MalformedPolicyDocumentException)

The request was rejected because the policy document was malformed. The error message describes the specific error.

PackedPolicyTooLargeException(PackedPolicyTooLargeException)

The request was rejected because the total packed size of the session policies and session tags combined was too large. An Amazon Web Services conversion compresses the session policy document, session policy ARNs, and session tags into a packed binary format that has a separate limit. The error message indicates by percentage how close the policies and tags are to the upper size limit. For more information, see Passing Session Tags in STS in the IAM User Guide.

You could receive this error even though you meet other defined session policy and session tag limits. For more information, see IAM and STS Entity Character Limits in the IAM User Guide.

RegionDisabledException(RegionDisabledException)

STS is not activated in the requested region for the account that is being asked to generate credentials. The account administrator must use the IAM console to activate STS in that region. For more information, see Activating and Deactivating Amazon Web Services STS in an Amazon Web Services Region in the IAM User Guide.

Unhandled(Unhandled)

๐Deprecated: Matching Unhandled directly is not forwards compatible. Instead, match using a variable wildcard pattern and check .code(): ยยยerr if err.code() == Some("SpecificExceptionCode") => { /* handle the error */ } See ProvideErrorMetadata for what information is available for the error.

An unexpected error occurred (e.g., invalid JSON returned by the service or an unknown error code).

Implementations

impl AssumeRoleWithSAMLError

pub fn unhandled(err: impl Into<Box<dyn Error + Send + Sync + 'static>>) -> Self

Creates the AssumeRoleWithSAMLError::Unhandled variant from any error type.

pub fn generic(err: ErrorMetadata) -> Self

Creates the AssumeRoleWithSAMLError::Unhandled variant from an ErrorMetadata.

pub fn meta(&self) -> &ErrorMetadata

Returns error metadata, which includes the error code, message, request ID, and potentially additional information.

pub fn is_expired_token_exception(&self) -> bool

Returns true if the error kind is AssumeRoleWithSAMLError::ExpiredTokenException.

pub fn is_idp_rejected_claim_exception(&self) -> bool

Returns true if the error kind is AssumeRoleWithSAMLError::IdpRejectedClaimException.

pub fn is_invalid_identity_token_exception(&self) -> bool

Returns true if the error kind is AssumeRoleWithSAMLError::InvalidIdentityTokenException.

pub fn is_malformed_policy_document_exception(&self) -> bool

Returns true if the error kind is AssumeRoleWithSAMLError::MalformedPolicyDocumentException.

pub fn is_packed_policy_too_large_exception(&self) -> bool

Returns true if the error kind is AssumeRoleWithSAMLError::PackedPolicyTooLargeException.

pub fn is_region_disabled_exception(&self) -> bool

Returns true if the error kind is AssumeRoleWithSAMLError::RegionDisabledException.