Module aws_sdk_sts::output

Output structures for operations.

Modules

assume_role_output

See AssumeRoleOutput

assume_role_with_saml_output

See AssumeRoleWithSamlOutput

assume_role_with_web_identity_output

See AssumeRoleWithWebIdentityOutput

decode_authorization_message_output

See DecodeAuthorizationMessageOutput

get_access_key_info_output

See GetAccessKeyInfoOutput

get_caller_identity_output

See GetCallerIdentityOutput

get_federation_token_output

See GetFederationTokenOutput

get_session_token_output

See GetSessionTokenOutput

Structs

AssumeRoleOutput

Contains the response to a successful AssumeRole request, including temporary Amazon Web Services credentials that can be used to make Amazon Web Services requests.

AssumeRoleWithSamlOutput

Contains the response to a successful AssumeRoleWithSAML request, including temporary Amazon Web Services credentials that can be used to make Amazon Web Services requests.

AssumeRoleWithWebIdentityOutput

Contains the response to a successful AssumeRoleWithWebIdentity request, including temporary Amazon Web Services credentials that can be used to make Amazon Web Services requests.

DecodeAuthorizationMessageOutput

A document that contains additional information about the authorization status of a request from an encoded message that is returned in response to an Amazon Web Services request.

GetAccessKeyInfoOutput

GetCallerIdentityOutput

Contains the response to a successful GetCallerIdentity request, including information about the entity making the request.

GetFederationTokenOutput

Contains the response to a successful GetFederationToken request, including temporary Amazon Web Services credentials that can be used to make Amazon Web Services requests.

GetSessionTokenOutput

Contains the response to a successful GetSessionToken request, including temporary Amazon Web Services credentials that can be used to make Amazon Web Services requests.