Enum aws_sdk_sts::Error

#[non_exhaustive]
pub enum Error {
    ExpiredTokenException(ExpiredTokenException),
    IdpCommunicationErrorException(IdpCommunicationErrorException),
    IdpRejectedClaimException(IdpRejectedClaimException),
    InvalidAuthorizationMessageException(InvalidAuthorizationMessageException),
    InvalidIdentityTokenException(InvalidIdentityTokenException),
    MalformedPolicyDocumentException(MalformedPolicyDocumentException),
    PackedPolicyTooLargeException(PackedPolicyTooLargeException),
    RegionDisabledException(RegionDisabledException),
    Unhandled(Unhandled),
}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ExpiredTokenException(ExpiredTokenException)

The web identity token that was passed is expired or is not valid. Get a new identity token from the identity provider and then retry the request.

IdpCommunicationErrorException(IdpCommunicationErrorException)

The request could not be fulfilled because the identity provider (IDP) that was asked to verify the incoming identity token could not be reached. This is often a transient error caused by network conditions. Retry the request a limited number of times so that you don't exceed the request rate. If the error persists, the identity provider might be down or not responding.

IdpRejectedClaimException(IdpRejectedClaimException)

The identity provider (IdP) reported that authentication failed. This might be because the claim is invalid.

If this error is returned for the AssumeRoleWithWebIdentity operation, it can also mean that the claim has expired or has been explicitly revoked.

InvalidAuthorizationMessageException(InvalidAuthorizationMessageException)

The error returned if the message passed to DecodeAuthorizationMessage was invalid. This can happen if the token contains invalid characters, such as linebreaks.

InvalidIdentityTokenException(InvalidIdentityTokenException)

The web identity token that was passed could not be validated by Amazon Web Services. Get a new identity token from the identity provider and then retry the request.

MalformedPolicyDocumentException(MalformedPolicyDocumentException)

The request was rejected because the policy document was malformed. The error message describes the specific error.

PackedPolicyTooLargeException(PackedPolicyTooLargeException)

The request was rejected because the total packed size of the session policies and session tags combined was too large. An Amazon Web Services conversion compresses the session policy document, session policy ARNs, and session tags into a packed binary format that has a separate limit. The error message indicates by percentage how close the policies and tags are to the upper size limit. For more information, see Passing Session Tags in STS in the IAM User Guide.

You could receive this error even though you meet other defined session policy and session tag limits. For more information, see IAM and STS Entity Character Limits in the IAM User Guide.

RegionDisabledException(RegionDisabledException)

STS is not activated in the requested region for the account that is being asked to generate credentials. The account administrator must use the IAM console to activate STS in that region. For more information, see Activating and Deactivating Amazon Web Services STS in an Amazon Web Services Region in the IAM User Guide.

Unhandled(Unhandled)

๐Deprecated: Matching Unhandled directly is not forwards compatible. Instead, match using a variable wildcard pattern and check .code(): ยยยerr if err.code() == Some("SpecificExceptionCode") => { /* handle the error */ } See ProvideErrorMetadata for what information is available for the error.

An unexpected error occurred (e.g., invalid JSON returned by the service or an unknown error code).