Module aws_sdk_sts::error

Errors that can occur when calling the service.

Modules

expired_token_exception

See ExpiredTokenException

idp_communication_error_exception

See IdpCommunicationErrorException

idp_rejected_claim_exception

See IdpRejectedClaimException

invalid_authorization_message_exception

See InvalidAuthorizationMessageException

invalid_identity_token_exception

See InvalidIdentityTokenException

malformed_policy_document_exception

See MalformedPolicyDocumentException

packed_policy_too_large_exception

See PackedPolicyTooLargeException

region_disabled_exception

See RegionDisabledException

Structs

AssumeRoleError

Error type for the AssumeRole operation.

AssumeRoleWithSAMLError

Error type for the AssumeRoleWithSAML operation.

AssumeRoleWithWebIdentityError

Error type for the AssumeRoleWithWebIdentity operation.

DecodeAuthorizationMessageError

Error type for the DecodeAuthorizationMessage operation.

ExpiredTokenException

The web identity token that was passed is expired or is not valid. Get a new identity token from the identity provider and then retry the request.

GetAccessKeyInfoError

Error type for the GetAccessKeyInfo operation.

GetCallerIdentityError

Error type for the GetCallerIdentity operation.

GetFederationTokenError

Error type for the GetFederationToken operation.

GetSessionTokenError

Error type for the GetSessionToken operation.

IdpCommunicationErrorException

The request could not be fulfilled because the identity provider (IDP) that was asked to verify the incoming identity token could not be reached. This is often a transient error caused by network conditions. Retry the request a limited number of times so that you don't exceed the request rate. If the error persists, the identity provider might be down or not responding.

IdpRejectedClaimException

The identity provider (IdP) reported that authentication failed. This might be because the claim is invalid.

InvalidAuthorizationMessageException

The error returned if the message passed to DecodeAuthorizationMessage was invalid. This can happen if the token contains invalid characters, such as linebreaks.

InvalidIdentityTokenException

The web identity token that was passed could not be validated by Amazon Web Services. Get a new identity token from the identity provider and then retry the request.

MalformedPolicyDocumentException

The request was rejected because the policy document was malformed. The error message describes the specific error.

PackedPolicyTooLargeException

The request was rejected because the total packed size of the session policies and session tags combined was too large. An Amazon Web Services conversion compresses the session policy document, session policy ARNs, and session tags into a packed binary format that has a separate limit. The error message indicates by percentage how close the policies and tags are to the upper size limit. For more information, see Passing Session Tags in STS in the IAM User Guide.

RegionDisabledException

STS is not activated in the requested region for the account that is being asked to generate credentials. The account administrator must use the IAM console to activate STS in that region. For more information, see Activating and Deactivating Amazon Web Services STS in an Amazon Web Services Region in the IAM User Guide.

Enums

AssumeRoleErrorKind

Types of errors that can occur for the AssumeRole operation.

AssumeRoleWithSAMLErrorKind

Types of errors that can occur for the AssumeRoleWithSAML operation.

AssumeRoleWithWebIdentityErrorKind

Types of errors that can occur for the AssumeRoleWithWebIdentity operation.

DecodeAuthorizationMessageErrorKind

Types of errors that can occur for the DecodeAuthorizationMessage operation.

GetAccessKeyInfoErrorKind

Types of errors that can occur for the GetAccessKeyInfo operation.

GetCallerIdentityErrorKind

Types of errors that can occur for the GetCallerIdentity operation.

GetFederationTokenErrorKind

Types of errors that can occur for the GetFederationToken operation.

GetSessionTokenErrorKind

Types of errors that can occur for the GetSessionToken operation.