aws_sdk_sts

Module types

Data structures used by operation inputs/outputs.

Modules

builders

Builders

error

Error types that AWS Security Token Service can respond with.

Structs

AssumedRoleUser

The identifiers for the temporary security credentials that the operation returns.

Credentials

Amazon Web Services credentials for API authentication.

FederatedUser

Identifiers for the federated user that is associated with the credentials.

PolicyDescriptorType

A reference to the IAM managed policy that is passed as a session policy for a role session or a federated user session.

ProvidedContext

Contains information about the provided context. This includes the signed and encrypted trusted context assertion and the context provider ARN from which the trusted context assertion was generated.

Tag

You can pass custom key-value pair attributes when you assume a role or federate a user. These are called session tags. You can then use the session tags to control access to resources. For more information, see Tagging Amazon Web Services STS Sessions in the IAM User Guide.